+++ addon/components/object-list-view.js
@@ -46,14 +46,6 @@ export default FlexberryBaseComponent.extend(
           });
         }).then(()=> {
           this.set('contentWithKeys', this.contentForRender);
-
-          // TODO: analyze this observers.
-          let attrsArray = this._getAttributesName();
-          content.forEach((record) => {
-            attrsArray.forEach((attrName) => {
-              Ember.addObserver(record, attrName, this, '_attributeChanged');
-            });
-          });
           this.set('showLoadingTbodyClass', false);
         });
       } else {
@@ -61,14 +53,6 @@ export default FlexberryBaseComponent.extend(
           this._addModel(item);
         });
         this.set('contentWithKeys', this.contentForRender);
-
-        // TODO: analyze this observers.
-        let attrsArray = this._getAttributesName();
-        content.forEach((record) => {
-          attrsArray.forEach((attrName) => {
-            Ember.addObserver(record, attrName, this, '_attributeChanged');
-          });
-        });
         this.set('showLoadingTbodyClass', false);
       }
     } else {
@@ -1080,15 +1064,6 @@ export default FlexberryBaseComponent.extend(
       this.get('objectlistviewEventsService').off('refreshList', this, this._refreshList);
 
       this._super(...arguments);
-
-      let content = this.get('content');
-      let attrsArray = this._getAttributesName();
-
-      content.forEach((record) => {
-        attrsArray.forEach((attrName) => {
-          Ember.removeObserver(record, attrName, this, '_attributeChanged');
-        });
-      });
     },
 
     /**
@@ -1665,11 +1640,6 @@ export default FlexberryBaseComponent.extend(
           this._addModel(modelToAdd);
           this.get('content').addObject(modelToAdd);
           this.get('objectlistviewEventsService').rowAddedTrigger(componentName, modelToAdd);
-
-          let attrsArray = this._getAttributesName();
-          attrsArray.forEach((attrName) => {
-            Ember.addObserver(modelToAdd, attrName, this, '_attributeChanged');
-          });
         }
       }
     },
@@ -1735,11 +1705,6 @@ export default FlexberryBaseComponent.extend(
 
       let componentName = this.get('componentName');
       this.get('objectlistviewEventsService').rowDeletedTrigger(componentName, record, immediately);
-
-      let attrsArray = this._getAttributesName();
-      attrsArray.forEach((attrName) => {
-        Ember.removeObserver(record, attrName, this, '_attributeChanged');
-      });
     },
 
     /**
@@ -1895,20 +1860,5 @@ export default FlexberryBaseComponent.extend(
       }
 
       return attrsArray;
-    },
-
-    /**
-    That observer is called when change attributes of model.
-
-    @method _attributeChanged
-    @private
-  */
-    _attributeChanged(record, attrName) {
-      let rowConfig = record.get('rowConfig');
-      let configurateRow = this.get('configurateRow');
-      if (configurateRow) {
-        Ember.assert('configurateRow must be a function', typeof configurateRow === 'function');
-        configurateRow(rowConfig, record);
-      }
     },
   });